+++ Integrations/MicrosoftGraphGroups/MicrosoftGraphGroups.py
@@ -116,7 +116,7 @@ class Client(BaseClient):
     Client to use in the MS Graph Groups integration. Overrides BaseClient
     """
     def __init__(self, base_url: str, tenant: str, auth_and_token_url: str, auth_id: str, token_retrieval_url: str,
-                 enc_key: str, verify: bool, proxy: dict):
+                 enc_key: str, verify: bool, proxy: bool):
         super().__init__(base_url, verify, proxy)
         self.tenant = tenant
         self.auth_and_token_url = auth_and_token_url
@@ -133,13 +133,21 @@ class Client(BaseClient):
         integration_context = demisto.getIntegrationContext()
         access_token = integration_context.get('access_token')
         valid_until = integration_context.get('valid_until')
+        calling_context = demisto.callingContext.get('context', {})  # type: ignore[attr-defined]
+        brand_name = calling_context.get('IntegrationBrand', '')
+        instance_name = calling_context.get('IntegrationInstance', '')
         if access_token and valid_until:
             if epoch_seconds() < valid_until:
                 return access_token
+
+        headers = {'Accept': 'application/json'}
+        headers['X-Content-Version'] = CONTENT_RELEASE_VERSION
+        headers['X-Branch-Name'] = CONTENT_BRANCH_NAME
+        headers['X-Content-Name'] = brand_name or instance_name or 'Name not found'
         try:
             dbot_response = requests.post(
                 self.token_retrieval_url,
-                headers={'Accept': 'application/json'},
+                headers=headers,
                 data=json.dumps({
                     'app_name': APP_NAME,
                     'registration_id': self.auth_id,
@@ -567,7 +575,7 @@ def main():
     auth_id = auth_and_token_url[0]
     enc_key = demisto.params().get('enc_key')
     verify = not demisto.params().get('insecure', False)
-    proxy = handle_proxy()
+    proxy = demisto.params().get('proxy') == 'true'
     if len(auth_and_token_url) != 2:
         token_retrieval_url = 'https://oproxy.demisto.ninja/obtain-token'  # guardrails-disable-line
     else: