+++ Integrations/RSANetWitness_v11_1/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
 
+## [19.12.1] - 2019-12-25
+  - Fixed an issue where environment proxy effects the integration when even though no proxy should be used.
+
+
 ## [19.11.0] - 2019-11-12
   - Added the *Fetch Limit* parameter.
   - Fixed an issue where an unsupported timestamp format caused the integration to fail.